+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -12,3 +12,4 @@ pub mod literals;
 pub mod modifiers;
 pub mod pragma_directives;
 pub mod source_units;
+pub mod statements;
