+++ aderyn_core/src/ast/mod.rs
@@ -3,7 +3,6 @@ pub mod impls;
 pub mod macros;
 pub mod magic;
 pub mod node_type;
-pub mod statements;
 pub mod structures;
 pub mod types;
 pub mod user_defined_value_types;
@@ -12,8 +11,8 @@ pub mod variables;
 pub mod yul;
 
 pub use self::{
-    impls::*, magic::*, node_type::*, statements::*, structures::*, types::*,
-    user_defined_value_types::*, using_for_directives::*, variables::*, yul::*,
+    impls::*, magic::*, node_type::*, structures::*, types::*, user_defined_value_types::*,
+    using_for_directives::*, variables::*, yul::*,
 };
 
 pub use ast_nodes::*;
